create resnet layer & remove resnet from util.py
